Back End Login
Implements in LoginController the login with an email and password in Firestone.
Login with google implements ok and when the login its correct his datas are added in to the firestone database.
Segue for go to Home View when the login its correct added but doesn't works because HomeController miss.
Missing save the user datas in to coredata, implements antoher type for login like facebook and the auto login with coredata